+++ hub/dataload/kg_loader.py
@@ -1,13 +1,7 @@
-import base64
-import math
-import zlib
-from collections import defaultdict
 from functools import partial
-from typing import Iterable
-
-import msgpack
 
 from hub.dataload.compressed_parser import load_from_tar
+from hub.dataload.info_parser import get_adj_list, encapsule, split_n_chunks
 from hub.dataload.utils.pipeline import apply_processors
 from hub.dataload.utils.process_category import process_category_list
 from hub.dataload.utils.process_node_fields import process_chembl_black_box_warning
@@ -54,52 +48,7 @@ def parser(*args, **kwargs):
     yield from map(processor_pipeline, load_from_tar(*args, **kwargs))
 
 
-
-
-def get_adj_list(edge_iterator: Iterable) -> dict[str, list[str]]:
-    adj_set = defaultdict(set[str])
-
-    for edge in edge_iterator:
-        sub = edge['subject']
-        obj = edge['object']
-        adj_set[sub].add(obj)
-
-    adj_list: dict[str, list[str]] = {
-        k: list(v) for k, v in adj_set.items()
-    }
-
-    return adj_list
-
-def encapsule(payload: list | dict) -> str:
-
-    # todo test pickle
-    payload_b = msgpack.packb(payload, use_bin_type=True)
-
-    # todo zst
-    compressed_b = zlib.compress(payload_b, level=6)
-    base64_payload = base64.b64encode(compressed_b).decode("ascii")
-
-    return base64_payload
-
-
-def split_n_chunks(b64: str, n: int, key: str, offset = 0) -> list:
-    if n <= 0:
-        raise ValueError("n must be > 0")
-
-    length = len(b64)
-    chunk_size = math.ceil(length / n)
-
-    return [
-        {
-            "_id": str(offset + i),
-            "key": key,
-            "chunk_index": i,
-            "value": b64[i : i + chunk_size]
-        } for i in range(0, length, chunk_size)
-    ]
-
-
-def flat_parser(*args, **kwargs):
+def node_info_parser(*args, **kwargs):
     result_key = kwargs.pop('result_key')
     adj_list_key = kwargs.pop('adj_list_key')
